requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Faker==36.1.1
Flask==3.0.3
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
kafka-python==2.0.3
MarkupSafe==3.0.2
narwhals==1.27.1
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
plotly==6.0.0
py4j==0.10.9.7
pyarrow==19.0.0
pyspark==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
retrying==1.3.4
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.0.6
zipp==3.21.0
